Bugfix: Peak values badly printed due to bad pointer typecast (double* to float*)

+++ inc/general_stats.h
@@ -1,6 +1,6 @@
 
 void global_average(float newsample, int infinite_counter);
 void detect_peak(float newsample);
-void detect_core_peak(float* newsample);
+void detect_core_peak(double* newsample);
 void report_dump(int sig);
 void init_stats();
